Move definition to own file
